+++ CHANGELOG.md
@@ -41,3 +41,7 @@
 - Found: step_1_data_preparation/step1.4/README.md
 - Found: step_1_data_preparation/step1.4/create_training_dataset.py
 - Found: step_1_data_preparation/step1.4/run_step1.4.sh
+
+## [2025-11-17 23:54:42] Auto Commit
+
+- Found: step_1_data_preparation/step1.5/